Scaffold GUHVI integration
