+++ process/configuration/regions/example_ES_Las_Palmas_2023.yml
@@ -498,5 +498,6 @@ reporting:
 ## Optional Google Earth Engine API
 ## Set 'gee' to true and enter your Google Cloud project ID to initate Google Earth Engine API connection and include the following indicators in your analysis:
 ## 1. Large Public Urban Green Space (LPUGS) accessibility and availability
+## 2. Global Urban Heat Vulnerability Index (GUHVI)
 gee: true
 gee_project_id: ee-global-indicators
